Backport recently merged karafka-rdkafka fixes
Brings rdkafka-ruby in sync with the fixes merged into karafka-rdkafka since the
last sync, preserving rdkafka-ruby's original error-handling style (no
RdkafkaError.build/build_from_c/validate!).

- Consumer GC finalizer now closes the consumer and destroys its consumer queue
- Free the librdkafka string in Consumer#cluster_id/#member_id + clusterid arity
- rd_kafka_query_watermark_offsets bound with blocking: true (releases the GVL)
- Destroy rd_kafka_conf_t on failed client creation (native_config/native_kafka)
- Metadata retry no longer leaks the native struct/topic per retried attempt
- Producer#produce removes the delivery handle on any post-register failure
- Delivery callback guarded with rescue + ensure-unlock
- Admin describe_configs/incremental_alter_configs/list_offsets parse args
  before allocating native resources (no leak on rejected input)
- to_native_tpl destroys the native list when population fails partway
- incremental_alter_configs raises on a rejected entry and frees the error object
- Admin result-count out-parameter allocated as :size_t instead of :int32

+++ lib/rdkafka/admin/describe_configs_report.rb
@@ -19,7 +19,7 @@ module Rdkafka
           .each do |config_resource_result_ptr|
             config_resource_result = ConfigResourceBindingResult.new(config_resource_result_ptr)
 
-            pointer_to_size_t = FFI::MemoryPointer.new(:int32)
+            pointer_to_size_t = FFI::MemoryPointer.new(:size_t)
             configs_ptr = Bindings.rd_kafka_ConfigResource_configs(
               config_resource_result_ptr,
               pointer_to_size_t
